backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
#fastapi-authlib
python-dotenv==1.0.0
jinja2
authlib==1.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6